+++ FeatureExtractor/FeatureExtractor/util.h
@@ -6,7 +6,7 @@
 #include <algorithm>
 #include "db.h"
 
-unsigned int levenshteinDistance(const std::string& s1, const std::string& s2);
+double levenshteinDistance(const std::string& s1, const std::string& s2);
 double jaroDistance(const std::string &s1, const std::string &s2);
 void generateSmallDomain(DB *db);
 void stringToLower(std::string& s);
